+++ src/App.js
@@ -2,7 +2,7 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 
 import Header from 'components/Common/Header/index';
 import MainPage from 'pages/MainPage';
-import Transfer from 'pages/Transfer';
+import TrainTransfer from 'pages/TrainTransfer/main';
 import TransferResult from 'pages/TransferResult';
 import BusSchedulePage from 'pages/BusSchedulePage';
 
@@ -14,7 +14,7 @@ function App() {
 					<Header />
 					<Routes>
 						<Route path='/' element={<MainPage />} />
-						<Route path='/train' element={<Transfer />} />
+						<Route path='/train' element={<TrainTransfer />} />
 						<Route path='/train/time-table' element={<TransferResult />} />
 						<Route path='/bus' element={<BusSchedulePage />} />
 					</Routes>
