+++ src/pages/TrainTransfer/components/TransferForm/FormStations.js
@@ -1,6 +1,6 @@
 import FormStationsName from './FormStationsName';
 import FormStationsHeader from './FormStationsHeader';
-import useFormStations from '../hooks/useFormStations';
+import useFormStations from '../../hooks/useFormStations';
 import FormStationsLists from './FormStationsLists';
 
 function FormStations({
